added validations

+++ services/todoService.js
@@ -1,19 +1,25 @@
 const uuid = require('uuid');
-// const { db } = require('../models/mongoose/todosModel');
 // const Todo = require('../models/mongoose/todosModel');
 const db = require('../models/sequelize/index')
+const TodoValidator = require('../validators/todoValidators')
 
 
 module.exports = class TodoService{
+
     static async createTodo(description) {
 
+        const { error, isValid } = await TodoValidator.todoCreation(description)
+
+        if(!isValid) {
+            return error
+        }
         // let newTodo = new Todo({
         //     "uniqueId" : uuid.v4(),
         //     "description": description,
         //     "isCompleted": false
         // })
 
-        // return newTodo.save();
+        //  return newTodo.save();
 
         return db.create({
             "uniqueId" : uuid.v4(),
@@ -33,6 +39,12 @@ module.exports = class TodoService{
     
 
     static async getTodoById(paramsId){
+
+        const { error, isValid } = await TodoValidator.retrieveTodoById(paramsId)
+
+        if(!isValid) {
+            return error
+        }
         // return Todo.findById(paramsId)
 
         return db.findOne({ where: {uniqueId: paramsId} })
@@ -40,6 +52,12 @@ module.exports = class TodoService{
     }
 
     static async removeTodo(paramsId) {
+
+        const { error, isValid } = await TodoValidator.retrieveTodoById(paramsId)
+
+        if(!isValid) {
+            return error
+        }
         // return Todo.findOneAndDelete({paramsId})
 
         return db.destroy({ where: {uniqueId: paramsId} })
@@ -47,6 +65,13 @@ module.exports = class TodoService{
     }
 
     static async updateTodo(description, paramsId) {
+
+        const { error, isValid } = await TodoValidator.retrieveTodoById(paramsId)
+
+        if(!isValid) {
+            return error
+        }
+
         return db.update({ description: description}, {where: {uniqueId: paramsId}})
 
     }
